Volume Action crash fix
Fixed an issue where the volume action card could cause a crash if it received a 'dual line message' from the receiver.
For a fast fix it uses the older style of args.device.log instead of the levelled log function; same as the other flow cards, so there's a fix available now. In future might replace those log functions but it's not really important for now.