V1.3.x
Following is concatenated commit messages going from 1.2.1 to 1.3.4. Note that some bug fixes mentioned are regarding the upgrade so did not fix existing issues in 1.2.1.

---
Bump to 1.3.0
There was an issue where the increased timeout time would trigger a connect failed timeout because socket setTimeout is for inactivity, not just for timeout during initial connect.
This has been replaced with a manual setTimeout, but needs more testing to see if it works correctly with a disconnected receiver.

Added findMessage function that will process multiline responses and looks for a reply we expect. So if a message also contains a volume response or some other thing we don't want, we just keep looking until a line matches what we expect. This fixes the issue where we expected a channel (SIGAME) but we got MVMAX for example from a previous request.
---
Added fix for settings only getting updated on restart of the app due to a change to Homey's onSettings callback.
--- 
Added change source to (specific source) card. Which uses a dropdown.
Removed SI prefix from source changed string token card.

Changed card titles to not use capital letters for each word, same as other built-in homey cards.
---
Added bug page url and support mailto.
---
Added CD as source option in the list.
---
Fixed bug in AND card that would cause everything to fail because of multiline messages not being filtered.
---
Somehow the turn on/off capability got messed up. It was a read request, instead of writeClose, which never returns anything so it got stuck waiting. Luckily only in the test version.
---
Fixed mute capability always toggling instead of specified muted or unmuted state.
---
Fixed old mute/unmute card to use new FindMessage system. It would would fail because of the carriage return.
---
Changed dutch translation to Ingangsbron instead of Kanaal or Source.
Fixed example saying MEDIAPLAYER instead of MPLAY. 

Increased slow poll mode failed attempts to 9 instead of 5, so two full status messages + 1.
---
Added (obsolete card) warning to all duplicate legacy cards. These might be actually be set to deprecated in a later release.
Don't want to stir too much shit at once.

Changed volume card titles to match the capability Homey ones. Removed the dB label from the variable and force it into the title itself because it won't show dB in the title when selecting the card from the menu. This makes it impossible to know the difference between the regular Homey volume and dB volume cards. Now you always see the dB label in all menus.
---
Undid JSON capability mess up. Relative volume works as expected again.
The issue it tried to fix, clamping the volume 0-1, has been implemented in the device.