Fail-safe added for socket inactivity
The previous version removed socket.setTimeout and replaced it with a manual timeout because socket.setTimeout also triggers on socket inactvity. Which we didn't want because of the new multi-response collecting.

But by removing it, if for any reason the socket is never closed, the app could get stuck waiting forever.
This version adds it back in the setTimeout, which triggers after 5 seconds of inactvity.